# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/book/create": {
            "post": {
                "description": "新增图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "新增图书",
                "parameters": [
                    {
                        "description": "create book request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CreateBookRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.CreateBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.CreateBookResponse"
                        }
                    }
                }
            }
        },
        "/admin/book/delete/{book_id}": {
            "delete": {
                "description": "删除图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "删除图书",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "delete book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteBookResponse"
                        }
                    }
                }
            }
        },
        "/admin/book/record/all": {
            "get": {
                "description": "所有用户所有借阅图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "所有用户所有借阅图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllBookRecordsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllBookRecordsResponse"
                        }
                    }
                }
            }
        },
        "/admin/book/record/borrow": {
            "get": {
                "description": "所有用户待还图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "所有用户待还图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllBorrowedBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllBorrowedBookResponse"
                        }
                    }
                }
            }
        },
        "/admin/book/record/return": {
            "get": {
                "description": "所有用户已还图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "所有用户已还图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllReturnedBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListAllReturnedBookResponse"
                        }
                    }
                }
            }
        },
        "/admin/book/update": {
            "post": {
                "description": "更改图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "更改图书信息",
                "parameters": [
                    {
                        "description": "update book request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateBookRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateBookResponse"
                        }
                    }
                }
            }
        },
...
        "/book/borrow/{book_id}": {
            "get": {
                "description": "借阅图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "借阅图书",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "borrow book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.BorrowBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.BorrowBookResponse"
                        }
                    }
                }
            }
        },
        "/book/list/{page}/{page_size}": {
            "get": {
                "description": "分页图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "图书列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBooksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBooksResponse"
                        }
                    }
                }
            }
        },
        "/book/record/all": {
            "get": {
                "description": "所有借阅图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "所有借阅图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBookRecordsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBookRecordsResponse"
                        }
                    }
                }
            }
        },
        "/book/record/borrow": {
            "get": {
                "description": "待还图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "待还图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBorrowedBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListBorrowedBookResponse"
                        }
                    }
                }
            }
        },
        "/book/record/return": {
            "get": {
                "description": "已还图书记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "已还图书记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListReturnedBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListReturnedBookResponse"
                        }
                    }
                }
            }
        },
        "/book/return/{book_id}": {
            "get": {
                "description": "归还图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "归还图书",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "return book id",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ReturnBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ReturnBookResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "contract.BorrowBookResponse": {
            "type": "object"
        },
        "contract.CreateBookRequest": {
            "type": "object"
        },
        "contract.CreateBookResponse": {
            "type": "object"
        },
        "contract.DeleteBookResponse": {
            "type": "object"
        },
...
        "contract.ListAllBookRecordsResponse": {
            "type": "object"
        },
        "contract.ListAllBorrowedBookResponse": {
            "type": "object"
        },
        "contract.ListAllReturnedBookResponse": {
            "type": "object"
        },
...
        "contract.ListBookRecordsResponse": {
            "type": "object"
        },
        "contract.ListBooksResponse": {
            "type": "object"
        },
        "contract.ListBorrowedBookResponse": {
            "type": "object"
        },
        "contract.ListReturnedBookResponse": {
            "type": "object"
        },
...
        "contract.ReturnBookResponse": {
            "type": "object"
        },
...
        "contract.UpdateBookRequest": {
            "type": "object"
        },
        "contract.UpdateBookResponse": {
            "type": "object"
        },
...
    }
}