# api/model/contract/book.go
package contract

type ListBooksRequest struct {
}

type ListBooksResponse struct {
}

type CreateBookRequest struct {
}

type CreateBookResponse struct {
}

type UpdateBookRequest struct {
}

type UpdateBookResponse struct {
}

type DeleteBookRequest struct {
}

type DeleteBookResponse struct {
}