# api/model/contract/rent.go
package contract

type BorrowBookRequest struct {
}

type BorrowBookResponse struct {
}

type ReturnBookRequest struct {
}

type ReturnBookResponse struct {
}

type ListBorrowedBookRequest struct {
}

type ListBorrowedBookResponse struct {
}

type ListReturnedBookRequest struct {
}

type ListReturnedBookResponse struct {
}

type ListBookRecordsRequest struct {
}

type ListBookRecordsResponse struct {
}

type ListAllBorrowedBookRequest struct {
}

type ListAllBorrowedBookResponse struct {
}

type ListAllReturnedBookRequest struct {
}

type ListAllReturnedBookResponse struct {
}

type ListAllBookRecordsRequest struct {
}

type ListAllBookRecordsResponse struct {
}